countdown element design implemented
