+++ android/src/main/java/io/parity/signer/ui/theme/Theme.kt
@@ -98,30 +98,3 @@ fun SignerNewTheme(
 	)
 }
 
-val Colors.fill30: Color
-	get() = if (isLight) Color(0x4D000000) else Color(0x4DFFFFFF)
-
-val Colors.fill24: Color
-	get() = if (isLight) Color(0x3D000000) else Color(0x3DFFFFFF)
-
-val Colors.fill18: Color
-	get() = if (isLight) Color(0x2E000000) else Color(0x2EFFFFFF)
-
-val Colors.fill12: Color
-	get() = if (isLight) Color(0x1F000000) else Color(0x1FFFFFFF)
-
-val Colors.fill6: Color
-	get() = if (isLight) Color(0x0F000000) else Color(0x0FFFFFFF)
-
-
-val Colors.backgroundPrimary: Color
-	get() = if (isLight) Color(0xFFFFFFFF) else Color(0xFF1E1E23)
-
-val Colors.backgroundSecondary: Color
-	get() = if (isLight) Color(0xFFFFFFFF) else Color(0xFF2C2C30)
-
-val Colors.textSecondary: Color
-	get() = if (isLight) Color(0xA8000000) else Color(0xB0FFFFFF)
-
-val Colors.textTertiary: Color
-	get() = if (isLight) Color(0x73000000) else Color(0x7AFFFFFF)
